Remove all Kafka code; put in simple example of using generated code
