+++ source/extensions/filters/network/kafka/code_generator/generated_header_h.j2
@@ -1,0 +1,18 @@
+#pragma once
+
+#include <memory>
+
+namespace Envoy {
+namespace Extensions {
+namespace NetworkFilters {
+namespace Kafka {
+
+class GeneratedSomething {
+	int foo() {
+		return 13;
+	}
+};
+
+typedef std::unique_ptr<GeneratedSomething> GeneratedSomethingPtr;
+
+}}}}
