--- source/extensions/filters/network/kafka/kafka_request_parser.cc
@@ -1,16 +1,0 @@
-#include "extensions/filters/network/kafka/kafka_request_parser.h"
-
-namespace Envoy {
-namespace Extensions {
-namespace NetworkFilters {
-namespace Kafka {
-
-ParseResponse RequestStartParser::parse(absl::string_view&) {
-  const RequestHeaderDeserializerPtr ptr = std::make_unique<RequestHeaderDeserializer>();
-  return ParseResponse::stillWaiting();
-}
-
-} // namespace Kafka
-} // namespace NetworkFilters
-} // namespace Extensions
-} // namespace Envoy
