+++ source/extensions/filters/network/kafka/normal_code.h
@@ -1,0 +1,15 @@
+#pragma once
+
+namespace Envoy {
+namespace Extensions {
+namespace NetworkFilters {
+namespace Kafka {
+
+class NormalCode {
+  int doSomething();
+};
+
+} // namespace Kafka
+} // namespace NetworkFilters
+} // namespace Extensions
+} // namespace Envoy
