--- source/extensions/filters/network/kafka/protocol_code_generator/request_parser.j2
@@ -1,16 +1,0 @@
-{#
-  Template for top-level structure representing a request in Kafka protocol (e.g. ProduceRequest).
-  Rendered templates for each request in Kafka protocol will be put into 'requests.h' file.
-
-  This template handles binding the top-level structure deserializer
-  (e.g. ProduceRequestV0Deserializer) with RequestParser.
-#}
-
-{% for version in complex_type.versions %}class {{ complex_type.name }}V{{ version }}Parser:
-  public RequestParser<{{ complex_type.name }}, {{ complex_type.name }}V{{ version }}Deserializer>
-{
-public:
-  {{ complex_type.name }}V{{ version }}Parser(RequestContextSharedPtr ctx) : RequestParser{ctx} {};
-};
-
-{% endfor %}
