+++ tools/spelling_dictionary.txt
@@ -384,9 +384,7 @@ dereferencing
 deregistered
 deserialization
 deserialize
-deserialized
 deserializer
-deserializers
 dest
 destructor
 destructors
@@ -580,7 +578,6 @@ parameterizing
 params
 paren
 parentid
-parsers
 pcall
 pcap
 pclose
@@ -704,7 +701,6 @@ templating
 templatize
 templatized
 templatizing
-testee
 th
 thru
 tm
